# docs/swagger.yaml
basePath: /
definitions:
...
  model.OrganizationMember:
    properties:
      createdAt:
        type: string
      deactivatedAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  model.RequestInviteAcceptance:
    properties:
      password:
        maxLength: 100
        minLength: 8
        type: string
      token:
        type: string
    required:
    - password
    - token
    type: object
  model.RequestInviteSend:
    properties:
      email:
        type: string
      name:
        type: string
      role:
        enum:
        - admin
        - member
        - owner
        - Admin
        - Member
        - Owner
        type: string
    required:
    - email
    - role
    type: object
  model.RequestInviteUpdate:
    properties:
      name:
        type: string
      role:
        type: string
    type: object
  model.RequestLogin:
    properties:
      email:
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
...
  repository.MemberInviteInfo:
    properties:
      acceptedAt:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      expiredAt:
        type: string
      id:
        type: string
      invitedBy:
        type: string
      name:
        type: string
      organizationId:
        type: string
      organizationName:
        description: organization name
        type: string
      role:
        description: invite role
        type: string
      status:
        description: invite status
        type: string
      updatedAt:
        type: string
    type: object
  repository.OrganizationMemberWithRole:
    properties:
      createdAt:
        type: string
      deactivatedAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
      updatedAt:
        type: string
    type: object
  service.JWT:
    properties:
      expAt:
        type: string
      issuedAt:
        type: string
      refreshToken:
        $ref: '#/definitions/service.RefreshTokenResponse'
      token:
        type: string
    type: object
  service.MemberCreateResponse:
    properties:
      authToken:
        $ref: '#/definitions/service.JWT'
      member:
        $ref: '#/definitions/repository.OrganizationMemberWithRole'
    type: object
  service.RefreshTokenResponse:
    properties:
      expAt:
        type: string
      token:
        type: string
    type: object
host: platform-api.string-api.xyz
info:
  contact:
    email: support@stringxyz.com
    name: Platform Management API Support
    url: http://string.xyz
  description: Platform Management API for managing Organizations, Platforms, API
    keys, etc.
  title: Platform Management API
  version: "1.0"
paths:
...
  /invites:
    get:
      parameters:
      - description: Organization ID
        in: header
        name: organizationId
        required: true
        type: string
      - description: Invite status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repository.MemberInviteInfo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: List invites
      tags:
      - Invite
    post:
      consumes:
      - application/json
      parameters:
      - description: Member ID
        in: header
        name: memberId
        required: true
        type: string
      - description: Organization ID
        in: header
        name: organizationId
        required: true
        type: string
      - description: Invite Send Request
        in: body
        name: RequestInviteSend
        required: true
        schema:
          $ref: '#/definitions/model.RequestInviteSend'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repository.MemberInviteInfo'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Send an invite
      tags:
      - Invite
  /invites/{id}:
    delete:
      parameters:
      - description: Member ID
        in: header
        name: memberId
        required: true
        type: string
      - description: Invite ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Revoke an invite
      tags:
      - Invite
    get:
      parameters:
      - description: Invite ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.MemberInviteInfo'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Get an invite
      tags:
      - Invite
    patch:
      consumes:
      - application/json
      parameters:
      - description: Member ID
        in: header
        name: memberId
        required: true
        type: string
      - description: Invite ID
        in: path
        name: id
        required: true
        type: string
      - description: Invite Update Request
        in: body
        name: RequestInviteUpdate
        required: true
        schema:
          $ref: '#/definitions/model.RequestInviteUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.MemberInviteInfo'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Update an invite
      tags:
      - Invite
    post:
      consumes:
      - application/json
      parameters:
      - description: Invite ID
        in: path
        name: id
        required: true
        type: string
      - description: Invite Acceptance Request
        in: body
        name: RequestInviteAcceptance
        required: true
        schema:
          $ref: '#/definitions/model.RequestInviteAcceptance'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OrganizationMember'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Accept an invite
      tags:
      - Invite
  /invites/{id}/resend:
    post:
      parameters:
      - description: Member ID
        in: header
        name: memberId
        required: true
        type: string
      - description: Invite ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.MemberInviteInfo'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Resend an invite
      tags:
      - Invite
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RequestLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.OrganizationMemberWithRole'
        "400":
          description: Bad Request
          schema:
            type: integer
        "401":
          description: Unauthorized
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      summary: Login
      tags:
      - Login
  /login/logout:
    post:
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - RefreshTokenAuth: []
      - RefreshBearerAuth: []
      summary: Logout
      tags:
      - Login
  /login/refresh:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.MemberCreateResponse'
        "401":
          description: Unauthorized
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - RefreshTokenAuth: []
      - RefreshBearerAuth: []
      summary: Refresh token
      tags:
      - Login
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    in: header
    name: StringAdminJWT
  RefreshBearerAuth:
    type: apiKey
    in: header
    name: StringAdminRefreshToken
  JWTAuth:
    type: apiKey
    in: cookie
    name: StringAdminJWT
  RefreshTokenAuth:
    type: apiKey
    in: cookie
    name: StringAdminRefreshToken
swagger: "2.0"